Add client-side process detail component with voting functionality
- Extract ProcessDetailClient.tsx from page.tsx for client-side rendering
- Implement comprehensive voting interface with wallet integration
- Add vote status tracking, proof generation, and result visualization
- Support multiple question types and real-time vote monitoring
- Include stepper UI for vote submission workflow
